+++ packages/graphql-codegen-cli/src/config.ts
@@ -2,7 +2,6 @@ import { cosmiconfig, defaultLoaders } from 'cosmiconfig';
 import { TypeScriptLoader } from 'cosmiconfig-typescript-loader';
 import { resolve } from 'path';
 import {
-  DetailedError,
   Types,
   Profiler,
   createProfiler,
@@ -156,8 +155,7 @@ export async function loadContext(configFilePath?: string): Promise<CodegenConte
 
   if (!result) {
     if (configFilePath) {
-      throw new DetailedError(
-        `Config ${configFilePath} does not exist`,
+      throw new Error(
         `
         Config ${configFilePath} does not exist.
 
@@ -168,18 +166,16 @@ export async function loadContext(configFilePath?: string): Promise<CodegenConte
       );
     }
 
-    throw new DetailedError(
-      `Unable to find Codegen config file!`,
-      `
+    throw new Error(
+      `Unable to find Codegen config file! \n
         Please make sure that you have a configuration file under the current directory!
       `
     );
   }
 
   if (result.isEmpty) {
-    throw new DetailedError(
-      `Found Codegen config file but it was empty!`,
-      `
+    throw new Error(
+      `Found Codegen config file but it was empty! \n
         Please make sure that you have a valid configuration file under the current directory!
       `
     );