+++ packages/graphql-codegen-cli/src/utils/cli-error.ts
@@ -1,6 +1,4 @@
-import { DetailedError } from '@graphql-codegen/plugin-helpers';
-
-type CompositeError = Error | DetailedError;
+type CompositeError = Error;
 type ListrError = Error & { errors: CompositeError[] };
 export function isListrError(err: Error & { name?: unknown; errors?: unknown }): err is ListrError {
   return err.name === 'ListrError' && Array.isArray(err.errors) && err.errors.length > 0;