+++ packages/plugins/other/visitor-plugin-common/src/enum-values.ts
@@ -1,6 +1,5 @@
 import { EnumValuesMap, ParsedEnumValuesMap } from './types.js';
 import { GraphQLSchema, isEnumType, GraphQLEnumType } from 'graphql';
-import { DetailedError } from '@graphql-codegen/plugin-helpers';
 import { parseMapper } from './mappers.js';
 
 function escapeString(str: string) {
@@ -37,9 +36,9 @@ export function parseEnumValues({
     const invalidMappings = Object.keys(mapOrStr).filter(gqlName => !allEnums.includes(gqlName));
 
     if (invalidMappings.length > 0) {
-      throw new DetailedError(
-        `Invalid 'enumValues' mapping!`,
-        `The following types does not exist in your GraphQL schema: ${invalidMappings.join(', ')}`
+      throw new Error(
+        `Invalid 'enumValues' mapping! \n
+        The following types does not exist in your GraphQL schema: ${invalidMappings.join(', ')}`
       );
     }
 
@@ -74,9 +73,9 @@ export function parseEnumValues({
           },
         };
       }
-      throw new DetailedError(
-        `Invalid "enumValues" configuration`,
-        `Enum "${gqlIdentifier}": expected string or object (with enum values mapping)`
+      throw new Error(
+        `Invalid "enumValues" configuration \n
+        Enum "${gqlIdentifier}": expected string or object (with enum values mapping)`
       );
     }, {} as ParsedEnumValuesMap);
   }